Add private method
