--- lib/test.rb
@@ -1,6 +1,0 @@
-class Scraper
-    def add(a,b)
-      a + b
-    end
-  end
-
